requirements.txt
flask>=2.0.0
python-dateutil>=2.8.0
werkzeug>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=0.19.0
